Adds support for adding memeber qualifier objects on-the-fly
Creating a new member qualifier object is now as easy as writing the
following expression:

(def-member-qualifier
  (matcher '(private static field-name field-value))
  ...)

In other words, the work done in this branch (i.e. experiment-20110220)
makes it very easy to define new semantics as part of class definitions
in the object system.

The grammar of 'def-member-qualifier' form will likely undergo changes in
the future; so will the name of the current abstractions and modules.

Acknowledgement
---------------
I'd  like to thank the Racket community of developers [1] for providing us
with such great platform for programming language experimentation.

This commit resolves the issue of compile-time state management using
a direct application of the pattern proposed in section 3.2.3 of the
Advanced Macrology article [2].

[1] http://racket-lang.org/
[2] http://is.gd/eqhV8b