Adds support for arbitrary handling of actions of class members in the object system
One such action is binding meta variables (aka. class variables.)

The following definition of a meta variable:

(public static foo 0)

Should appear as follows within the closure that class definitions ought to
expand to:

(define (class-name ...)
  (define (foo 0) ; we call this the binder action
  ...)

For now, actions are represented as lambdas that receive the original
syntax object representing the member definition i.e.:

The binder action will then glean the necessary information from the
syntax object and generates a list. That list will be placed in the
template that generates the closure that class definitions expand to.