Attempts to add literal-ids to the generated syntax-case
The goal is to be able to distinguish class members based on
the content and order of their qualifiers.

Literal identifiers should be declared as such so that
qualifiers such as public and static won't be bound to template
variables.

We also need to do so to be able to distinguish between e.g.
public and private members. Should they be bound as template
variables, then a template variable named public or private (*)
will be created. This variable will then be bound to either
public or private regardless of the content of the original
syntax object, leaving public and private fields
undistinguished.

(*) Whether the template variable's name is public or private
depending on the order in which syntax-case clauses are
generated.