Simplifies and _corrects_ the semantics of the objective system by grouping members based on their scope
We now preserve the order in which class members are declared
and only make a distinction between class members and instance
members.

This because, instance members are allowed to see class members,
but NOT the other way around.

Prior to this change, we used to e.g. first bind public class
fields and methods, then private class fields and methods, etc.